+++ src/test/java/org/jabref/logic/bst/TestBstVM.java
@@ -1,46 +1,63 @@
 package org.jabref.logic.bst;
 
 import java.io.IOException;
-import java.io.StringReader;
 import java.nio.file.Path;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.Stack;
 
-import org.jabref.logic.importer.ImportFormatPreferences;
-import org.jabref.logic.importer.ParserResult;
-import org.jabref.logic.importer.fileformat.BibtexParser;
 import org.jabref.model.entry.BibEntry;
-import org.jabref.model.util.DummyFileUpdateMonitor;
+import org.jabref.model.entry.field.StandardField;
+import org.jabref.model.entry.types.StandardEntryType;
 
 import org.antlr.v4.runtime.CharStreams;
 import org.antlr.v4.runtime.RecognitionException;
 import org.junit.jupiter.api.Test;
-import org.mockito.Answers;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.Mockito.mock;
 
 public class TestBstVM {
+
+    public static BibEntry defaultTestEntry() {
+        /*
+            @inproceedings{canh05,
+            author = {Crowston, K. and Annabi, H. and Howison, J. and Masango, C.},
+            title = {Effective work practices for floss development: A model and propositions},
+            booktitle = {Hawaii International Conference On System Sciences (HICSS)},
+            year = {2005},
+            owner = {oezbek},
+            timestamp = {2006.05.29},
+            url = {http://james.howison.name/publications.html}}
+        */
+
+        return new BibEntry(StandardEntryType.InProceedings)
+                .withCitationKey("canh05")
+                .withField(StandardField.AUTHOR, "Crowston, K. and Annabi, H. and Howison, J. and Masango, C.")
+                .withField(StandardField.TITLE, "Effective work practices for floss development: A model and propositions")
+                .withField(StandardField.BOOKTITLE, "Hawaii International Conference On System Sciences (HICSS)")
+                .withField(StandardField.YEAR, "2005")
+                .withField(StandardField.OWNER, "oezbek")
+                .withField(StandardField.TIMESTAMP, "2006.05.29")
+                .withField(StandardField.URL, "http://james.howison.name/publications.html");
+    }
 
     @Test
     public void testAbbrv() throws RecognitionException, IOException {
         TestVM vm = new TestVM(Path.of("src/test/resources/org/jabref/logic/bst/abbrv.bst"));
-        List<BibEntry> v = List.of(t1BibtexEntry());
+        List<BibEntry> testEntries = List.of(defaultTestEntry());
 
         String expected = "\\begin{thebibliography}{1}\\bibitem{canh05}K.~Crowston, H.~Annabi, J.~Howison, and C.~Masango.\\newblock Effective work practices for floss development: A model and  propositions.\\newblock In {\\em Hawaii International Conference On System Sciences (HICSS)}, 2005.\\end{thebibliography}";
+        String result = vm.render(testEntries);
 
-        assertEquals(expected.replaceAll("\\s", ""), vm.render(v).replaceAll("\\s", ""));
+        assertEquals(
+                expected.replaceAll("\\s", ""),
+                result.replaceAll("\\s", ""));
     }
 
     @Test
-    public void testSimple() throws RecognitionException, IOException {
+    public void testSimple() throws RecognitionException {
         TestVM vm = new TestVM("""
                ENTRY { address author title type } { } { label }
                INTEGERS { output.state before.all mid.sentence after.sentence after.block }
@@ -53,9 +70,9 @@ public class TestBstVM {
                STRINGS { s t }
                READ
                """);
-        List<BibEntry> v = List.of(t1BibtexEntry());
+        List<BibEntry> testEntries = List.of(defaultTestEntry());
 
-        vm.render(v);
+        vm.render(testEntries);
 
         assertEquals(2, vm.getStrings().size());
         assertEquals(7, vm.getIntegers().size());
@@ -65,20 +82,23 @@ public class TestBstVM {
     }
 
     @Test
-    public void testLabel() throws RecognitionException, IOException {
+    public void testLabel() throws RecognitionException {
         TestVM vm = new TestVM("""
                 ENTRY { title } {} { label }
-                FUNCTION { test } { label #0 = title 'label := #5 label #6 pop$ }
+                FUNCTION { test } {
+                    label #0 =
+                    title 'label :=
+                    #5 label #6 pop$ }
                 READ
                 ITERATE { test }
                 """);
-        List<BibEntry> v = List.of(t1BibtexEntry());
+        List<BibEntry> testEntries = List.of(defaultTestEntry());
 
-        vm.render(v);
+        vm.render(testEntries);
 
-        assertEquals("Effective work practices for floss development: A model and propositions", vm
-                .getStack()
-                .pop());
+        assertEquals(
+                "Effective work practices for floss development: A model and propositions",
+                vm.getStack().pop());
     }
 
     @Test
@@ -87,381 +107,6 @@ public class TestBstVM {
 
         vm.render(Collections.emptyList());
         assertEquals("\"\"", vm.getStack().pop());
-    }
-
-    @Test
-    public void testFunction1() throws RecognitionException {
-        TestVM vm = new TestVM("FUNCTION { init.state.consts }{ #0 'before.all := } ");
-
-        vm.render(Collections.emptyList());
-
-        assertEquals(38, vm.getFunctions().size());
-        assertNotNull(vm.getFunctions().get("init.state.consts"));
-    }
-
-    @Test
-    public void testExecuteSimple() throws RecognitionException {
-        TestVM vm = new TestVM("""
-                INTEGERS { variable.a }
-                FUNCTION { init.state.consts }{ #5 'variable.a := }
-                EXECUTE { init.state.consts }
-                """);
-
-        vm.render(Collections.emptyList());
-
-        assertEquals(Integer.valueOf(5), vm.getIntegers().get("variable.a"));
-    }
-
-    @Test
-    public void testExecuteSimple2() throws RecognitionException {
-        TestVM vm = new TestVM("""
-                FUNCTION { a } {
-                    #5 #5 =
-                    #1 #2 =
-                    #3 #4 <
-                    #4 #3 <
-                    #4 #4 <
-                    #3 #4 >
-                    #4 #3 >
-                    #4 #4 >
-                    "H" "H" =
-                    "H" "Ha" =
-                }
-                EXECUTE { a }
-                """);
-
-        vm.render(Collections.emptyList());
-
-        assertEquals(BstVM.FALSE, vm.getStack().pop());
-        assertEquals(BstVM.TRUE, vm.getStack().pop());
-        assertEquals(BstVM.FALSE, vm.getStack().pop());
-        assertEquals(BstVM.TRUE, vm.getStack().pop());
-        assertEquals(BstVM.FALSE, vm.getStack().pop());
-        assertEquals(BstVM.FALSE, vm.getStack().pop());
-        assertEquals(BstVM.FALSE, vm.getStack().pop());
-        assertEquals(BstVM.TRUE, vm.getStack().pop());
-        assertEquals(BstVM.FALSE, vm.getStack().pop());
-        assertEquals(BstVM.TRUE, vm.getStack().pop());
-        assertEquals(0, vm.getStack().size());
-    }
-
-    @Test
-    public void testIfSkipPop() throws RecognitionException {
-        TestVM vm = new TestVM("""
-                FUNCTION { not } { { #0 } { #1 } if$ }
-                FUNCTION { and } { 'skip$ { pop$ #0 } if$ }
-                FUNCTION { or } { { pop$ #1 } 'skip$ if$ }
-                FUNCTION { test } {
-                    #1 #1 and
-                    #0 #1 and
-                    #1 #0 and
-                    #0 #0 and
-                    #0 not
-                    #1 not
-                    #1 #1 or
-                    #0 #1 or
-                    #1 #0 or
-                    #0 #0 or
-                }
-                EXECUTE { test }
-                """);
-
-        vm.render(Collections.emptyList());
-
-        assertEquals(BstVM.FALSE, vm.getStack().pop());
-        assertEquals(BstVM.TRUE, vm.getStack().pop());
-        assertEquals(BstVM.TRUE, vm.getStack().pop());
-        assertEquals(BstVM.TRUE, vm.getStack().pop());
-        assertEquals(BstVM.FALSE, vm.getStack().pop());
-        assertEquals(BstVM.TRUE, vm.getStack().pop());
-        assertEquals(BstVM.FALSE, vm.getStack().pop());
-        assertEquals(BstVM.FALSE, vm.getStack().pop());
-        assertEquals(BstVM.FALSE, vm.getStack().pop());
-        assertEquals(BstVM.TRUE, vm.getStack().pop());
-        assertEquals(0, vm.getStack().size());
-    }
-
-    @Test
-    public void testArithmetic() throws RecognitionException {
-        TestVM vm = new TestVM("""
-                FUNCTION { test } {
-                    #1 #1 +
-                    #5 #2 -
-                }
-                EXECUTE { test }
-                """);
-
-        vm.render(Collections.emptyList());
-
-        assertEquals(3, vm.getStack().pop());
-        assertEquals(2, vm.getStack().pop());
-        assertEquals(0, vm.getStack().size());
-    }
-
-    @Test
-    public void testArithmetic2() throws RecognitionException {
-        TestVM vm = new TestVM("""
-                FUNCTION { test } {
-                    #1 "HELLO" +      % Throws exception
-                    #5 #2 -
-                }
-                EXECUTE { test }
-                """);
-
-        assertThrows(BstVMException.class, () -> vm.render(Collections.emptyList()));
-    }
-
-    @Test
-    public void testNumNames() throws RecognitionException {
-        TestVM vm = new TestVM("""
-                FUNCTION { test } { "Johnny Foo and Mary Bar" num.names$ }
-                EXECUTE { test }
-                """);
-
-        vm.render(Collections.emptyList());
-
-        assertEquals(2, vm.getStack().pop());
-        assertEquals(0, vm.getStack().size());
-    }
-
-    @Test
-    public void testNumNames2() throws RecognitionException {
-        TestVM vm = new TestVM("""
-                FUNCTION { test } { "Johnny Foo { and } Mary Bar" num.names$ }
-                EXECUTE { test }
-                """);
-
-        vm.render(Collections.emptyList());
-
-        assertEquals(1, vm.getStack().pop());
-        assertEquals(0, vm.getStack().size());
-    }
-
-    @Test
-    public void testStringOps1() throws RecognitionException {
-        TestVM vm = new TestVM("""
-                FUNCTION { test } {
-                    "H" "allo" * "Johnny" add.period$ "Johnny." add.period$
-                    "Johnny!" add.period$ "Johnny?" add.period$ "Johnny} }}}" add.period$
-                    "Johnny!}" add.period$ "Johnny?}" add.period$ "Johnny.}" add.period$
-                }
-                EXECUTE { test }
-                """);
-
-        vm.render(Collections.emptyList());
-
-        assertEquals("Johnny.}", vm.getStack().pop());
-        assertEquals("Johnny?}", vm.getStack().pop());
-        assertEquals("Johnny!}", vm.getStack().pop());
-        assertEquals("Johnny.}", vm.getStack().pop());
-        assertEquals("Johnny?", vm.getStack().pop());
-        assertEquals("Johnny!", vm.getStack().pop());
-        assertEquals("Johnny.", vm.getStack().pop());
-        assertEquals("Johnny.", vm.getStack().pop());
-        assertEquals("Hallo", vm.getStack().pop());
-        assertEquals(0, vm.getStack().size());
-    }
-
-    @Test
-    public void testSubstring() throws RecognitionException {
-        TestVM vm = new TestVM("""
-                FUNCTION { test } {
-                    "123456789" #2  #1          substring$  % 2
-                    "123456789" #4  global.max$ substring$  % 456789
-                    "123456789" #1  #9          substring$  % 123456789
-                    "123456789" #1  #10         substring$  % 123456789
-                    "123456789" #1  #99         substring$  % 123456789
-                    "123456789" #-7 #3          substring$  % 123
-                    "123456789" #-1 #1          substring$  % 9
-                    "123456789" #-1 #3          substring$  % 789
-                    "123456789" #-2 #2          substring$  % 78
-                } EXECUTE { test }
-                """);
-
-        vm.render(Collections.emptyList());
-
-        assertEquals("78", vm.getStack().pop());
-        assertEquals("789", vm.getStack().pop());
-        assertEquals("9", vm.getStack().pop());
-        assertEquals("123", vm.getStack().pop());
-        assertEquals("123456789", vm.getStack().pop());
-        assertEquals("123456789", vm.getStack().pop());
-        assertEquals("123456789", vm.getStack().pop());
-        assertEquals("456789", vm.getStack().pop());
-        assertEquals("2", vm.getStack().pop());
-        assertEquals(0, vm.getStack().size());
-    }
-
-    @Test
-    public void testEmpty() throws RecognitionException, IOException {
-        TestVM vm = new TestVM("""
-                ENTRY { title } { } { }
-                READ
-                STRINGS { s }
-                FUNCTION { test } {
-                    s empty$          % TRUE
-                    "" empty$         % TRUE
-                    "   " empty$      % TRUE
-                    title empty$      % TRUE
-                    " HALLO " empty$  % FALSE
-                }
-                ITERATE { test }
-                """);
-        List<BibEntry> v = List.of(TestBstVM.bibtexString2BibtexEntry("@article{a, author=\"AAA\"}"));
-
-        vm.render(v);
-
-        assertEquals(BstVM.FALSE, vm.getStack().pop());
-        assertEquals(BstVM.TRUE, vm.getStack().pop());
-        assertEquals(BstVM.TRUE, vm.getStack().pop());
-        assertEquals(BstVM.TRUE, vm.getStack().pop());
-        assertEquals(BstVM.TRUE, vm.getStack().pop());
-        assertEquals(0, vm.getStack().size());
-    }
-
-    @Test
-    public void testDuplicateEmptyPopSwapIf() throws RecognitionException {
-        TestVM vm = new TestVM("""
-                FUNCTION { emphasize } {
-                    duplicate$ empty$
-                        { pop$ "" }
-                        { "{\\em " swap$ * "}" * }
-                    if$
-                }
-                FUNCTION { test } {
-                    "" emphasize
-                    "Hello" emphasize
-                }
-                EXECUTE { test }
-                """);
-
-        vm.render(Collections.emptyList());
-
-        assertEquals("{\\em Hello}", vm.getStack().pop());
-        assertEquals("", vm.getStack().pop());
-        assertEquals(0, vm.getStack().size());
-    }
-
-    @Test
-    public void testChangeCase() throws RecognitionException {
-        TestVM vm = new TestVM("""
-                STRINGS { title }
-                READ
-                FUNCTION { format.title } {
-                    duplicate$ empty$
-                        { pop$ "" }
-                        { "t" change.case$ }
-                    if$
-                }
-                FUNCTION { test } {
-                    "hello world" "u" change.case$ format.title
-                    "Hello World" format.title
-                    "" format.title
-                    "{A}{D}/{C}ycle: {I}{B}{M}'s {F}ramework for {A}pplication {D}evelopment and {C}ase" "u" change.case$ format.title
-                }
-                EXECUTE { test }
-                """);
-
-        vm.render(Collections.emptyList());
-
-        assertEquals("{A}{D}/{C}ycle: {I}{B}{M}'s {F}ramework for {A}pplication {D}evelopment and {C}ase",
-                vm.getStack().pop());
-        assertEquals("", vm.getStack().pop());
-        assertEquals("Hello world", vm.getStack().pop());
-        assertEquals("Hello world", vm.getStack().pop());
-        assertEquals(0, vm.getStack().size());
-    }
-
-    @Test
-    public void testTextLength() throws RecognitionException {
-        TestVM vm = new TestVM("""
-                FUNCTION { test } {
-                    "hello world" text.length$
-                    "Hello {W}orld" text.length$
-                    "" text.length$
-                    "{A}{D}/{Cycle}" text.length$
-                    "{\\This is one character}" text.length$
-                    "{\\This {is} {one} {c{h}}aracter as well}" text.length$
-                    "{\\And this too" text.length$
-                    "These are {\\11}" text.length$
-                }
-                EXECUTE { test }
-                """);
-
-        vm.render(Collections.emptyList());
-
-        assertEquals(11, vm.getStack().pop());
-        assertEquals(1, vm.getStack().pop());
-        assertEquals(1, vm.getStack().pop());
-        assertEquals(1, vm.getStack().pop());
-        assertEquals(8, vm.getStack().pop());
-        assertEquals(0, vm.getStack().pop());
-        assertEquals(11, vm.getStack().pop());
-        assertEquals(11, vm.getStack().pop());
-        assertEquals(0, vm.getStack().size());
-    }
-
-    @Test
-    public void testIntToStr() throws RecognitionException {
-        TestVM vm = new TestVM("""
-                FUNCTION { test } { #3 int.to.str$ #9999 int.to.str$ }
-                EXECUTE { test }
-                """);
-
-        vm.render(Collections.emptyList());
-
-        assertEquals("9999", vm.getStack().pop());
-        assertEquals("3", vm.getStack().pop());
-        assertEquals(0, vm.getStack().size());
-    }
-
-    @Test
-    public void testChrToInt() throws RecognitionException {
-        TestVM vm = new TestVM("""
-                FUNCTION { test } { "H" chr.to.int$ }
-                EXECUTE { test }
-                """);
-
-        vm.render(Collections.emptyList());
-
-        assertEquals(72, vm.getStack().pop());
-        assertEquals(0, vm.getStack().size());
-    }
-
-    @Test
-    public void testChrToIntIntToChr() throws RecognitionException {
-        TestVM vm = new TestVM("""
-                FUNCTION { test } { "H" chr.to.int$ int.to.chr$ }
-                EXECUTE {test}
-                """);
-
-        vm.render(Collections.emptyList());
-
-        assertEquals("H", vm.getStack().pop());
-        assertEquals(0, vm.getStack().size());
-    }
-
-    @Test
-    public void testSort() throws RecognitionException, IOException {
-        TestVM vm = new TestVM("""
-                ENTRY { title } { } { label }
-                FUNCTION { presort } { cite$ 'sort.key$ := }
-                ITERATE { presort }
-                SORT
-                """);
-        List<BibEntry> v = List.of(
-                TestBstVM.bibtexString2BibtexEntry("@article{a, author=\"AAA\"}"),
-                TestBstVM.bibtexString2BibtexEntry("@article{b, author=\"BBB\"}"),
-                TestBstVM.bibtexString2BibtexEntry("@article{d, author=\"DDD\"}"),
-                TestBstVM.bibtexString2BibtexEntry("@article{c, author=\"CCC\"}"));
-
-        vm.render(v);
-
-        List<BstEntry> v2 = vm.getEntries();
-        assertEquals(Optional.of("a"), v2.get(0).entry.getCitationKey());
-        assertEquals(Optional.of("b"), v2.get(1).entry.getCitationKey());
-        assertEquals(Optional.of("c"), v2.get(2).entry.getCitationKey());
-        assertEquals(Optional.of("d"), v2.get(3).entry.getCitationKey());
     }
 
     @Test
@@ -494,273 +139,17 @@ public class TestBstVM {
     }
 
     @Test
-    public void testWhile() throws RecognitionException {
-        TestVM vm = new TestVM("""
-                        STRINGS { t }
-                        FUNCTION {not} { { #0 } { #1 } if$ }
-                        FUNCTION {n.dashify} {
-                            "HELLO-WORLD" 't :=
-                            ""
-                            { t empty$ not }                                    % while
-                            {
-                                t #1 #1 substring$ "-" =                        % if
-                                {
-                                    t #1 #2 substring$ "--" = not               % if
-                                    {
-                                        "--" *
-                                        t #2 global.max$ substring$ 't :=
-                                    }
-                                    {
-                                        { t #1 #1 substring$ "-" = }            % while
-                                        {
-                                            "-" *
-                                            t #2 global.max$ substring$ 't :=
-                                        }
-                                        while$
-                                    }
-                                   if$
-                                }
-                                {
-                                    t #1 #1 substring$ *
-                                    t #2 global.max$ substring$ 't :=
-                                }
-                                if$
-                            }
-                            while$
-                        }
-                        EXECUTE {n.dashify}
-                        """);
-        List<BibEntry> v = Collections.emptyList();
-
-        vm.render(v);
-
-        assertEquals(1, vm.getStack().size());
-        assertEquals("HELLO--WORLD", vm.getStack().pop());
-    }
-
-    @Test
-    public void testType() throws RecognitionException, IOException {
-        TestVM vm = new TestVM("""
-                ENTRY { title }  { }  { label }
-                FUNCTION { presort } { cite$ 'sort.key$ := }
-                ITERATE { presort }
-                SORT
-                FUNCTION { test } { type$ }
-                ITERATE { test }
-                """);
-        List<BibEntry> v = List.of(
-                TestBstVM.bibtexString2BibtexEntry("@article{a, author=\"AAA\"}"),
-                TestBstVM.bibtexString2BibtexEntry("@book{b, author=\"BBB\"}"),
-                TestBstVM.bibtexString2BibtexEntry("@misc{c, author=\"CCC\"}"),
-                TestBstVM.bibtexString2BibtexEntry("@inproceedings{d, author=\"DDD\"}"));
-
-        vm.render(v);
-
-        assertEquals(4, vm.getStack().size());
-        assertEquals("inproceedings", vm.getStack().pop());
-        assertEquals("misc", vm.getStack().pop());
-        assertEquals("book", vm.getStack().pop());
-        assertEquals("article", vm.getStack().pop());
-    }
-
-    @Test
-    public void testMissing() throws RecognitionException, IOException {
-        TestVM vm = new TestVM("""
-                ENTRY { title }  { }  { label }
-                FUNCTION { presort } { cite$ 'sort.key$ := }
-                ITERATE { presort }
-                READ
-                SORT
-                FUNCTION { test } { title missing$ cite$ }
-                ITERATE { test }
-                """);
-        List<BibEntry> v = List.of(
-                t1BibtexEntry(),
-                TestBstVM.bibtexString2BibtexEntry("@article{test, author=\"No title\"}"));
-
-        vm.render(v);
-
-        assertEquals(4, vm.getStack().size());
-        assertEquals("test", vm.getStack().pop());
-        assertEquals(1, vm.getStack().pop());
-        assertEquals("canh05", vm.getStack().pop());
-        assertEquals(0, vm.getStack().pop());
-    }
-
-    @Test
-    public void testFormatName() throws RecognitionException {
-        TestVM vm = new TestVM("""
-                FUNCTION { format }{ "Charles Louis Xavier Joseph de la Vall{\\'e}e Poussin" #1 "{vv~}{ll}{, jj}{, f}?" format.name$ }
-                EXECUTE { format }
-                """);
-        List<BibEntry> v = Collections.emptyList();
-
-        vm.render(v);
-
-        assertEquals("de~la Vall{\\'e}e~Poussin, C.~L. X.~J?", vm.getStack().pop());
-        assertEquals(0, vm.getStack().size());
-    }
-
-    @Test
-    public void testFormatName2() throws RecognitionException, IOException {
-        TestVM vm = new TestVM("""
-                ENTRY { author } { } { label }
-                FUNCTION { presort } { cite$ 'sort.key$ := }
-                ITERATE { presort }
-                READ
-                SORT
-                FUNCTION { format }{ author #2 "{vv~}{ll}{, jj}{, f}?" format.name$ }
-                ITERATE { format }
-                """);
-
-        List<BibEntry> v = List.of(
-                t1BibtexEntry(),
-                TestBstVM.bibtexString2BibtexEntry(
-                        "@book{test, author=\"Jonathan Meyer and Charles Louis Xavier Joseph de la Vall{\\'e}e Poussin\"}"));
-        vm.render(v);
-        assertEquals("de~la Vall{\\'e}e~Poussin, C.~L. X.~J?", vm.getStack().pop());
-        assertEquals("Annabi, H?", vm.getStack().pop());
-        assertEquals(0, vm.getStack().size());
-    }
-
-    @Test
-    public void testCallType() throws RecognitionException, IOException {
-        TestVM vm = new TestVM("""
-                ENTRY { title }  { }  { label }
-                FUNCTION { presort } { cite$ 'sort.key$ := }
-                ITERATE { presort }
-                READ
-                SORT
-                FUNCTION { inproceedings }{ "InProceedings called on " title * }
-                FUNCTION { book }{ "Book called on " title * }
-                ITERATE { call.type$ }
-                """);
-        List<BibEntry> v = List.of(
-                t1BibtexEntry(),
-                TestBstVM.bibtexString2BibtexEntry("@book{test, title=\"Test\"}"));
-
-        vm.render(v);
-
-        assertEquals(2, vm.getStack().size());
-        assertEquals("Book called on Test", vm.getStack().pop());
-        assertEquals(
-                "InProceedings called on Effective work practices for floss development: A model and propositions",
-                vm.getStack().pop());
-        assertEquals(0, vm.getStack().size());
-    }
-
-    @Test
-    public void testIterate() throws RecognitionException, IOException {
-        TestVM vm = new TestVM("""
-                ENTRY { address author title type } {} { label }
-                FUNCTION { test }{ cite$ }
-                READ
-                ITERATE { test }
-                """);
-        List<BibEntry> v = List.of(
-                t1BibtexEntry(),
-                TestBstVM.bibtexString2BibtexEntry("@article{test, title=\"BLA\"}"));
-
-        vm.render(v);
-
-        assertEquals(2, vm.getStack().size());
-        String s1 = (String) vm.getStack().pop();
-        String s2 = (String) vm.getStack().pop();
-
-        if ("canh05".equals(s1)) {
-            assertEquals("test", s2);
-        } else {
-            assertEquals("canh05", s2);
-            assertEquals("test", s1);
-        }
-    }
-
-    @Test
-    public void testWidth() throws RecognitionException, IOException {
-        TestVM vm = new TestVM("""
-                ENTRY { address author title type } { } { label }
-                STRINGS { longest.label }
-                INTEGERS { number.label longest.label.width }
-                FUNCTION { initialize.longest.label } {
-                    "" 'longest.label :=
-                    #1 'number.label :=
-                    #0 'longest.label.width :=
-                }
-                FUNCTION {longest.label.pass} {
-                    number.label int.to.str$ 'label :=
-                    number.label #1 + 'number.label :=
-                    label width$ longest.label.width >
-                        {
-                            label 'longest.label :=
-                            label width$ 'longest.label.width :=
-                        }
-                        'skip$
-                    if$
-                }
-                EXECUTE { initialize.longest.label }
-                ITERATE { longest.label.pass }
-                FUNCTION { begin.bib } {
-                    preamble$ empty$
-                        'skip$
-                        { preamble$ write$ newline$ }
-                    if$
-                    "\\begin{thebibliography}{"  longest.label  * "}" *
-                }
-                EXECUTE {begin.bib}
-                """);
-
-        List<BibEntry> v = List.of(t1BibtexEntry());
-
-        vm.render(v);
-
-        assertTrue(vm.getIntegers().containsKey("longest.label.width"));
-        assertEquals("\\begin{thebibliography}{1}", vm.getStack().pop());
-    }
-
-    @Test
-    public void testSwap() throws RecognitionException {
-        TestVM vm = new TestVM("""
-                FUNCTION { a } { #3 "Hallo" swap$ }
-                EXECUTE { a }
-                """);
-
-        List<BibEntry> v = Collections.emptyList();
-        vm.render(v);
-
-        assertEquals(2, vm.getStack().size());
-        assertEquals(3, vm.getStack().pop());
-        assertEquals("Hallo", vm.getStack().pop());
-    }
-
-    @Test
     public void testHypthenatedName() throws RecognitionException, IOException {
         TestVM vm = new TestVM(Path.of("src/test/resources/org/jabref/logic/bst/abbrv.bst"));
-        List<BibEntry> v = List.of(TestBstVM.bibtexString2BibtexEntry("@article{canh05, author = \"Jean-Paul Sartre\" }"));
-        assertTrue(vm.render(v).contains("J.-P. Sartre"));
-    }
+        List<BibEntry> testEntries = List.of(
+                new BibEntry(StandardEntryType.Article)
+                        .withCitationKey("canh05")
+                        .withField(StandardField.AUTHOR, "Jean-Paul Sartre")
+        );
 
-    private static BibEntry bibtexString2BibtexEntry(String s) throws IOException {
-        ParserResult result = new BibtexParser(mock(ImportFormatPreferences.class, Answers.RETURNS_DEEP_STUBS), new DummyFileUpdateMonitor()).parse(new StringReader(s));
-        Collection<BibEntry> c = result.getDatabase().getEntries();
-        assertEquals(1, c.size());
-        return c.iterator().next();
-    }
+        String result = vm.render(testEntries);
 
-    private static String t1BibtexString() {
-        return """
-                @inproceedings{canh05,
-                  author = {Crowston, K. and Annabi, H. and Howison, J. and Masango, C.},
-                  title = {Effective work practices for floss development: A model and propositions},
-                  booktitle = {Hawaii International Conference On System Sciences (HICSS)},
-                  year = {2005},
-                  owner = {oezbek},
-                  timestamp = {2006.05.29},
-                  url = {http://james.howison.name/publications.html}}
-                """;
-    }
-
-    public static BibEntry t1BibtexEntry() throws IOException {
-        return TestBstVM.bibtexString2BibtexEntry(t1BibtexString());
+        assertTrue(result.contains("J.-P. Sartre"));
     }
 
     public static class TestVM extends BstVM {
